--- src/components/header/group/index.ts
@@ -1,5 +1,0 @@
-export { default as CreateGroups } from './addGroupModal/AddGroupModal';
-export { default as EditGroupMembers } from './editGroupMembers/EditGroupMembers';
-export { default as EditGroupName } from './editGroupName/EditGroupName';
-export { default as GroupMenuButton } from './GroupMenuButton';
-export { default as SwitchEntity } from './switchEntity/SwitchEntity';
