+++ src/components/header/switchEntity/modules/listItem/groupListItem/GroupListItem.module.scss
@@ -1,0 +1,25 @@
+@import "../../../../../../assets/variable";
+
+.listItem {
+  display: flex;
+  align-items: center;
+  padding: 7px 12px;
+  cursor: pointer;
+
+  &:hover {
+    background-color: $color-blackSqueeze;
+  }
+}
+
+.checkIcon {
+  margin-right: 8px;
+}
+
+.transparencyIcon {
+  opacity: 0;
+}
+
+.name{
+  width: calc(100% - 30px);
+  margin-right: 16px;
+}
