+++ src/components/header/switchEntity/modules/modal/addGroupModal/add-group-modal.scss
@@ -1,6 +1,6 @@
-@import "../../../../assets/variable";
-@import "../../../../assets/mixin/icon-btn";
-@import "../../../../assets/mixin/operation-btn";
+@import "../../../../../../assets/variable";
+@import "../../../../../../assets/mixin/icon-btn";
+@import "../../../../../../assets/mixin/operation-btn";
 
 .add-group-modal {
   width: 400px;
