+++ src/components/header/switchEntity/modules/modal/editGroupMembers/edit-group-members.scss
@@ -1,4 +1,4 @@
-@import "../../../../assets/variable";
+@import "../../../../../../assets/variable";
 
 .edit-group-members {
   width: 400px;
