+++ src/components/header/switchEntity/modules/modal/editGroupName/edit-group-name.scss
@@ -1,4 +1,4 @@
-@import "../../../../assets/variable";
+@import "../../../../../../assets/variable";
 
 .edit-group-name {
   width: 400px;
