+++ src/containers/header/HeaderContainer.tsx
@@ -1,12 +1,10 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { useLocation, useParams } from 'react-router';
 import { push } from 'connected-react-router';
-import axios from 'axios';
 import { Groups } from '../../reducks/groups/types';
 import { getApprovedGroups } from '../../reducks/groups/selectors';
-import { fetchGroups } from '../../reducks/groups/operations';
-import { fetchUserInfo, logOut } from '../../reducks/users/operations';
+import { logOut } from '../../reducks/users/operations';
 import Header from '../../components/header/Header';
 
 const HeaderContainer = () => {
@@ -15,47 +13,6 @@ const HeaderContainer = () => {
   const currentPath = useLocation().pathname;
   const pathName = useLocation().pathname.split('/')[1];
   const approvedGroups: Groups = useSelector(getApprovedGroups);
-  const [name, setName] = useState<string>('');
-
-  useEffect(() => {
-    let unmount = false;
-    const signal = axios.CancelToken.source();
-    const setUserName = async () => {
-      await dispatch(fetchUserInfo(signal));
-      if (pathName !== 'group' && name === '' && !unmount) {
-        setName('グループ選択なし');
-      }
-    };
-    setUserName();
-    return () => {
-      unmount = true;
-      signal.cancel();
-    };
-  }, [pathName]);
-
-  useEffect(() => {
-    let unmount = false;
-    const signal = axios.CancelToken.source();
-    const setGroupName = async () => {
-      if (pathName === 'group' && name === '') {
-        await dispatch(fetchGroups(signal));
-        let groupName = '';
-        for (const group of approvedGroups) {
-          if (group.group_id === Number(group_id)) {
-            groupName = group.group_name;
-          }
-        }
-        if (!unmount) {
-          setName(groupName);
-        }
-      }
-    };
-    setGroupName();
-    return () => {
-      unmount = true;
-      signal.cancel();
-    };
-  }, [approvedGroups, pathName, group_id]);
 
   const existsGroupWhenRouting = (path: string) => {
     if (pathName !== 'group') {
@@ -98,8 +55,6 @@ const HeaderContainer = () => {
     <Header
       group_id={Number(group_id)}
       pathName={pathName}
-      name={name}
-      setName={setName}
       approvedGroups={approvedGroups}
       currentPage={currentPage}
       logOutCheck={logOutCheck}
