+++ src/containers/header/switchEntity/modules/button/GrouopMenuButtonContainer.tsx
@@ -1,10 +1,10 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { useDispatch } from 'react-redux';
 import { push } from 'connected-react-router';
-import { Group } from '../../../reducks/groups/types';
-import { groupWithdrawal } from '../../../reducks/groups/operations';
-import { getPathGroupId } from '../../../lib/path';
-import GroupMenuButton from '../../../components/header/group/GroupMenuButton';
+import { Group } from '../../../../../reducks/groups/types';
+import { groupWithdrawal } from '../../../../../reducks/groups/operations';
+import { getPathGroupId } from '../../../../../lib/path';
+import GroupMenuButton from '../../../../../components/header/switchEntity/modules/button/groupMenuButton/GroupMenuButton';
 
 interface GroupMenuButtonContainerProps {
   approvedGroup: Group;
@@ -15,9 +15,9 @@ const GroupMenuButtonContainer = (props: GroupMenuButtonContainerProps) => {
   const dispatch = useDispatch();
   const pathGroupId = getPathGroupId(window.location.pathname);
   const groupId = props.approvedGroup.group_id;
-  const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null);
+  const [anchorEl, setAnchorEl] = useState<null | SVGElement>(null);
 
-  const openGroupMenu = (event: React.MouseEvent<HTMLButtonElement>) => {
+  const openGroupMenu = (event: React.MouseEvent<SVGElement>) => {
     setAnchorEl(event.currentTarget);
   };
